Finally! The bug is GONE!!!!!
I spent the last 5 days trying to get rid of all the bugs that appeared after introducing the Legal moves generator, which lead to major changes happening to the rest of the code, mostly getting rid of useless parts of the code, like the board Dictionary in the Board class, or that we check if a move is legal twice in the Make move method. Also, I ditched Unity's resource intensive Instantiate and Destroy methods and instead I'm now using object pooling to instantiate and destroy pieces, which made the code run way faster, and got rid of some bugs in the Bot script that the Destroy method takes a few frames to destroy the piece. There are still some minor errors that occur when playing in Bot vs Bot mode, but i couldn't recreate them in the other Player modes, so, I won't be fixing them until i find a way to recreate them. Anyways, i'm pretty sure that we're finally ready to implement the first "AI" scripts that will allow the Bot to chose moves instead of playing one at random. Also, the rest of the week is going to be holidays, so I probably won't make any changes to the code at that period, so, i guess i'll see you next week🥳🥳🥳